+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 0.1.2 (2014-0X-XX)
 ------------------
 
-- Added: ``ServerError`` and ``ValidationError`` exceptions
+- Added: ``ServerError``, ``ValidationError`` and ``ResourceNoFieldsProvidedError`` exceptions
 - Added: Create functionality for User resource
 - Added: Informal string representation, i.e. __str__(), to all resources
 - Changed: reimplemented fix for __repr__() from 0.1.1